criação do jogosAdapter

+++ app/src/main/java/com/innaval/recycleview2/model/Jogo.kt
@@ -3,6 +3,6 @@ package com.innaval.recycleview2.model
 data class Jogo(
     var codigo:Int,
     var nomeJogo:String,
-    var notaJogo: Int,
+    var notaJogo: Float,
     var console:String
 )
